requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
typer>=0.12.0
rich>=13.7.0
pydantic>=2.6.0
click>=8.1.7
tomli-w>=1.0.0

# Additional modern dependencies for type hints and better async support
typing-extensions>=4.10.0
httpx>=0.27.0

# Azure SDK for deployed model discovery
azure-identity>=1.15.0
azure-mgmt-cognitiveservices>=13.5.0
azure-mgmt-resource>=23.1.0

# Testing dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0